Deformation gradient implementation

+++ src/velocities.h
@@ -15,7 +15,7 @@ double find_magnitude(const array<double, DIM>&);
 
 double find_velocity(const double&, const int&);
 
-double find_velocity_derivative(const double&, const double&);
+double find_velocity_derivative(const double&, const int&);
 
 double f(const double&);
